enigma2: 4.3.1r6 - fixed DVB-T(2) namespace problem introduced in 4.3.1r5 - fixed satfinder crash on change from predefined to manuel transponder   introduced in 4.3.1r5 - allow to input DVB-S(2) frequencies with one decimal place   needed to lock transponders with very low symbol rates

+++ usr/lib/enigma2/python/Components/TuneTest.py
@@ -16,7 +16,7 @@ class Tuner:
 		if self.frontend:
 			print "tuning to transponder with data", transponder
 			parm = eDVBFrontendParametersSatellite()
-			parm.frequency = transponder[0] * 1000
+			parm.frequency = int(transponder[0] * 1000)
 			parm.symbol_rate = transponder[1] * 1000
 			parm.polarisation = transponder[2]
 			parm.fec = transponder[3]
